clg_proj_JRS(backend)/backend/job-recommendation-server/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via
    #   nltk
    #   pip-tools
django==4.2.5
    # via
    #   -r requirements/requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-jazzmin
    #   djangorestframework
    #   drf-yasg
django-cors-headers==4.2.0
    # via -r requirements/requirements.in
django-filter==23.3
    # via -r requirements/requirements.in
django-jazzmin==2.6.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements/requirements.in
exceptiongroup==1.2.0
    # via pytest
factory-boy==3.3.0
    # via -r requirements/requirements.in
faker==24.1.0
    # via factory-boy
inflection==0.5.1
    # via drf-yasg
iniconfig==2.0.0
    # via pytest
joblib==1.3.2
    # via nltk
nltk==3.8.1
    # via -r requirements/requirements.in
numpy==1.26.4
    # via -r requirements/requirements.in
packaging==23.2
    # via
    #   build
    #   drf-yasg
    #   pytest
pillow==10.3.0
    # via -r requirements/requirements.in
pip-tools==7.3.0
    # via -r requirements/requirements.in
pluggy==1.4.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements/requirements.in
pyproject-hooks==1.0.0
    # via build
pytest==8.1.1
    # via
    #   -r requirements/requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via faker
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
regex==2023.12.25
    # via nltk
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
tqdm==4.66.2
    # via nltk
typing-extensions==4.8.0
    # via asgiref
uritemplate==4.1.1
    # via drf-yasg
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
